*** empty log message ***
git-svn-id: https://svn.java.net/svn/electric~svn/trunk@3564 b64ed6c1-75a8-e629-b1da-d0dae74a743a

+++ srcj/com/sun/electric/tool/user/CircuitChanges.java
@@ -1485,6 +1485,9 @@ public class CircuitChanges
 						if (eobj instanceof Export)
 						{
 							Export pp = (Export)eobj;
+							int errCode = cantEdit(cell, pp.getOriginalPort().getNodeInst(), true);
+							if (errCode < 0) return false;
+							if (errCode > 0) continue;
 							pp.kill();
 						}
 					}
